Criação das matrículas

+++ db/migrate/20200918175528_create_enrollments.rb
@@ -1,7 +1,7 @@
 class CreateEnrollments < ActiveRecord::Migration[6.0]
   def change
     create_table :enrollments do |t|
-      t.float :valor_total
+      t.decimal :valor_total
       t.integer :quant_faturas
       t.integer :dia_vencimento
       t.text :curso
